platform: msm: Clear BAM pipe pointer after USB disconnect
When the USB port is disocnnected from the host, all the BAM pipes are
closed and pipe data structure is freed. However, the pointer to this
structure has remained with a stale value. This has led to crashes after
re-connection since this pointer has been referenced again. Setting the
pointer value to NULL resolves this since there are NULL value checks
before this pointer is accessed.

Change-Id: I71db45d795f813b1f65af37756b9fe570b9b93d0
CRs-fixed: 576789
Signed-off-by: Danny Segal <dsegal@codeaurora.org>
Signed-off-by: ykong <ykong@sierrawireless.com>