USB: RMNET: smd_ctrl: Don't drop ctrl pkts if channel not open
RMNET control packets are dropped if SMD channel is not yet open.
Host would not be notified of the same and may recover only after
some timeout or retry mechanism. Fix this by queuing the control
packets when channel is not open and send them once is gets opened.

CRs-fixed: 627287
Change-Id: Ieb803c0ed6c5b1cd17178a59e7dcc6ae0c24b215
Signed-off-by: Manu Gautam <mgautam@codeaurora.org>
Signed-off-by: ykong <ykong@sierrawireless.com>