UBI: Dispatch update notification if the volume is updated
The UBI_IOCVOLUP ioctl is used to start an update and also to
truncate a volume. In the first case, a "volume updated" notification
is dispatched when the update is done.

This commit adds the "volume updated" notification to be also sent when
the volume is truncated. This is required for UBI block and gluebi to get
notified about the new volume size.

Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Acked-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Cc: stable@vger.kernel.org # v3.15+
(cherry picked from commit fda322a1b3b9e8ee231913c500f73c6988b1aff5)

Change-Id: I41513da2b13fe7706587f1bec135075cd124b082