UBI: return on error in rename_volumes()
I noticed this during a code review.  We are checking that the strlen()
of ->name is not less than the ->name_len which the user gave us.  I
believe this bug is harmless but clearly we meant to return here instead
of setting an error code and then not using it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Acked-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(cherry picked from commit 7fbbd05799976c0611dcb229649260504b2bdef5)

Change-Id: I18689c1d8a0e8d601281c32c188cd1f2b595f0c4