UBI: Fastmap: Ensure that only one fastmap work is scheduled
If the WL pool runs out of PEBs we schedule a fastmap write
to refill it as soon as possible.
Ensure that only one at a time is scheduled otherwise we might end in
a fastmap write storm because writing the fastmap can schedule another
write if bitflips are detected.

Signed-off-by: Richard Weinberger <richard@nod.at>
Reviewed-by: Tanya Brokhman <tlinder@codeaurora.org>
Reviewed-by: Guido Martínez <guido@vanguardiasur.com.ar>
Acked-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(cherry picked from commit 19371d73c9bd31a8e634ec5a80fc19fcd7714481)

Change-Id: I282007abf26df0fe8536eee1e8c5846734b7912a