UBI: Fastmap: Remove bogus ubi_assert()
It is legal to have PEBs left in the used list.
This can happen if UBI copies a PEB and a powercut happens
between writing a new fastmap and adding this PEB into the EBA table.
In this case the old PEB will be used.

Signed-off-by: Richard Weinberger <richard@nod.at>
Acked-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(cherry picked from commit a83832a7c8d023f446ec865926190951bd18a4b1)

Change-Id: I69750e1111a35f14c08ee220eea23925eb6af2ad