UBI: Fastmap: Remove eba_orphans logic
This logic is in vain as we treat protected PEBs also as used, so this
case must not happen.
If a PEB is found which is in the EBA table but not known as used
has to be issued as fatal error.

Signed-off-by: Richard Weinberger <richard@nod.at>
Acked-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
(cherry picked from commit d141a8ef21ab496ab6f8d188dfe7ca33e16fe798)

Change-Id: I028affa0defc32dcf3c026ebfbb178d690ab5501