msm_otg: Eliminated pm_suspend_work, USB link wake lock
This commit completely eliminates triggering of system suspend from
msm_otg's runtime suspend routine. System suspend is now triggered
by kernel's auto-sleep function and there is no more need to
trigger it from this driver (which was wrong by design as system
sleep shouldn't be a driver's decision).

Wake lock that was previously used to prevent sleep when USB link
is up is now re-commissioned to protect the state machine. This
frees the PHY's wake lock for other purposes. The implementation
paves the way to using EPOLLWAKEUP, with the intention of passing
the association between link state and system sleep to user-space,
where it could be made application-dependent.

This effectively reverts the following commit:
  ticket:1514 redo msm_otg: suspend/resume on cable unplug/plug
as requested in its FIXME note.

Change-Id: I4c32857dc96e04478bd0fa96dd00c07b45793109
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>