ehci-msm-hsic: Use internal wakeup_source
The external wake_lock structure was mis-used by releasing it in
_suspend() routine and grabbing it in _resume(). This prevented
the system to enter auto-sleep.

Convert the driver to use the device's internal wakeup_source
instead of adding another to its private block. Use wakeup_source
appropriately so as not to prevent the system from auto-sleep and
allow a timeout interval for user-space to process a device event.

Change-Id: I53f1bec95fb95fe7c0227847507f410fb6422486
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>