gpio_wake_n: greatly simplified driver
Implementation of gpio_wake_n driver was incongruent with the power
management framework of 3.4 kernel:
(a) it never allowed the system to auto-sleep
(b) it made explicit calls to pm_suspend().

Rather than fixing the previous implementation, pretty much rewrite
the driver to perform the following:
1. Configure wake-n_gpio for input, rising and falling interrupt edge.
2. In _probe() function, check status of wake-n_gpio and acquire/
   release wakeup_source accordingly.
3. In interrupt routine, do the same.

It was unclear why the driver had to implement any power management
routines, so these were removed completely. On the same token, it
was unclear why the driver needed the XO resource, so this code was
nuked as well.

Change-Id: I227d7075b8d1783c4a4a40a08254739fba35979a
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>