Smack: Rework file hooks
This is one of those cases where you look at code you did
years ago and wonder what you might have been thinking.
There are a number of LSM hooks that work off of file pointers,
and most of them really want the security data from the inode.
Some, however, really want the security context that the process
had when the file was opened. The difference went undetected in
Smack until it started getting used in a real system with real
testing. At that point it was clear that something was amiss.

This patch corrects the misuse of the f_security value in several
of the hooks. The behavior will not usually be any different, as
the process had to be able to open the file in the first place, and
the old check almost always succeeded, as will the new, but for
different reasons.

Thanks to the Samsung Tizen development team that identified this.

Dragan: I've removed smack_file_set_fowner change in order to avoid
changes introduced by commit e0b93eddfe17dcb7d644eb5d6ad02a86fc41a977:

security: make security_file_set_fowner, f_setown and __f_setown
    void return

Resolves: Jira:LXSWI9X1517-123
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
  (cherry picked from commit 5e7270a6dd14fa6e3bb10128f200305b4a75f350)
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: Id6fd0eeed3023d5e09640823f56812c558eb47bd