arm: gic: Configure the GIC to run in secure mode
commit 26e44869e1e730ec7434e899dfd5857530b63415 from
git://codeaurora.org/quic/le/kernel/msm msm-3.4

Configure the GIC to run in secure mode and handle
secure as well as non-secure interrupts. This patch
adds an API to configure an IRQ as a secure IRQ so that
it can be treated as an FIQ in the secure mode.

Change-Id: Ic3321e76c95a4c10d6287ba418e84623e7004cb1
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
Signed-off-by: Sathish Ambley <sambley@codeaurora.org>
Signed-off-by: Weiwei Wang <weiwei.wang@windriver.com>

Signed-off-by: Mihai Dragusu <mihai.dragusu@windriver.com>