trac:2285 trac:1950 msm_otg: notify userland on DISCONNECT
Boards that internally drive VBUS high mistakenly interpreted a
DISCONNECT event as a SUSPEND event. Assuming that the interface
was suspended, USB gadgets attempted to perform a USB remote
resume of host and failed because the host was not present.

When VBUS is driven low on disconnect, the OTG interface enters
IDLE state (as opposed to SUSPEND). USB gadgets are disconnected
and hence cannot remote-resume the host. This event, however, was
not reported to userland and the system stayed awake due to "link
active" wake lock.

Reorganize link status notification to report whenever USB OTG is
suspended/resumed, as opposed to when the host requests suspend or
resume. This change is backwards-compatible with boards that drive
VBUS high, although the issue of mis-interpreting DISCONNECT as
SUSPEND persists due to previous hardware design.

patch2: disabled VDDMIN workaround on non-device-bus suspend

Change-Id: I8a6133351cac36a138609745bd0835e678a16e16
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>