ticket:2506 smd_dev: Keep smd interrupts active during suspend
The default setting for smd_dev interrupts was to disable them
during suspend. The default processing flow for this type of
interrupts is to mark them as 'pending' and go back to sleep.
These interrupts would then be processed when the system wakes up
for another reason.

This fixes the issue where AT commands weren't fully processed
when issued from a UART AT console device, which in some cases
caused the modem watchdog to bark. When the AT console is a USB
ACM device, the issued is masked by the system staying awake due
to USB connection.

Add IRQF_NO_SUSPEND to default smd_dev IRQ flags.

Change-Id: I17cd8ad40df1798f6a13abbe0e8a3681bd14b859
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>