msm_hsic: Restart controller after suspend/resume cycle
EHCI code seems to rely on the fact that the last controller
command would always be cached in the ehci_hcd struct. Somehow
this was not the case for Qualcomm's HSIC controller: after a
suspend/resume cycle, the cached value did not capture the last
CMD_RUN issued after ports were resumed. As a result, when the
next command was issued based on its cached value, the controller
was halted. This caused the attached USB device to fail its resume
with error -110.

Re-read the value of USB command register in the last stages of
msm_hsic_bus_resume.

Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>