msm_sdcc: Fix recursive check of active detect work
MMC host resume triggers a rescan work, but the Qualcomm MMC
driver checked for an active detect work in its runtime suspend
function, which was part of the msmsdcc_disable() call. This
caused a recursion in which msmsdcc_disable() always failed with
a warning.

Skip checking for detect work active during runtime suspend. Since
this is all called from a hosts release procedure, we no longer
have to resume and re-suspend the host, and runtime_resume() can
unconditionally return zero.

Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>