SBM:12325 ANO:81784 - USB becomes unrecognized during long term selective suspend test
Based on case CR01299427 and CR02072651.

The code will basically cover any case where we get a very fast suspend/resume sequence from the host.
If we allow the device to go into PM suspend, the latency to handle the resume will be very long
(or even lost, because IRQs may be temporarily disabled while the system enters PM suspend).

Change-Id: I1e7b769a5576087523a693fdc39f9445393925a0
Signed-off-by: ykong <ykong@sierrawireless.com>
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>