Smack: IPv6 host labeling
IPv6 appears to be (finally) coming of age with the
influx of autonomous devices. In support of this, add
the ability to associate a Smack label with IPv6 addresses.

This patch also cleans up some of the conditional
compilation associated with the introduction of
secmark processing. It's now more obvious which bit
of code goes with which feature.

Resolves: Jira:LXSWI9X1517-123
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: Iee304206932aac3d64c4f581a1fd76a898fd1696