trac:2966: Fix backwards compatibility with YAFFS2 in kernel 3.4.91.
Currently, YAFFS2 tags have two parts: tags part and tags ECC part.
Tags part is 16 bytes, and ECC part is 12 bytes long. If 2K page flash
part is used (WP85), OOB area is 16 bytes long. It is obvious that full
tags structure could not fit in there.

However, YAFFS2 in Yocto 1.6 kernel (3.4.91), does not check for OOB
size, therefore allowing for OOB area overflow. Even though this is a
bug, we need to be backwards compatible, and need to emulate previous
YAFFS2 behaviour. As long as we are not using ECC on YAFFS2 tags, we
should be OK.

To be compatible with current YAFFS2 implementation, specify
'force-oobsize-check' mount option when mounting YAFFS2 partition.

Change-Id: I06ef9b42c51a061892a2dde513e7d5cd469e57dc