CUS80562-USB Device, Address State, wrong GET_STATUS(ep!=0) response
Problem Description:
Ci13xxx udc driver does not check the status and return the ep status for GET STATUG request.
This is conflict with USB 2.0 Spec. In chapter 9.4.5, if an interface or endpoint other than
endpoint zero is specified in address state, the device responds with a Request Error(STALL PID).

Solution:
Check the udc status, if not configuried, return stall PID as Request Error.

Change-Id: I94804ec0f5423a312af642dc154484f60a0f889e
Signed-off-by: Yoking Wang <yowang@sierrawireless.com>