trac:3959 - Prevent invalid I2C addressing via direct device access
Proper way of using I2C devices is to first set their address via IOCTL
call, and only then attempt read/write operation on them. However, there
is nothing to stop users to use I2C devices directly (e.g. cat <device>).

To make things worse, fp->private_data may look correct, because it is
coming directly from userland (e.g. it may not be initialized to NULL).
That leaves us with the few choices:
* Check client address
* Check if client really exists by probing it

In this case, we are going to rely on the fact that client address is
initialized to 0 (invalid for 7-bit I2C addressing). If this proves to be
ineffective, we may need to probe the client as well (the big hammer). At
this time, this looks unnecessary, but the code is there (#if 0-ed), as a
reminder that we may need to do it down the road.

Change-Id: I79d1c367a628e5ec9c12c079cd014074bdd39bfc
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>