perf: duplicate deletion of perf event
a malicious app can open a perf event with constraint_duplicate
bit set, disable the event, and close the fd.  On closing the fd,
the perf_release() modification causes the kernel to clean up
the event as if it still were enabled, leading to the event
being removed from a list twice.

Note (DM): struct perf_event_attr.constraint_duplicate member does
not exist, and its check was removed from the patch. __reserved_1
member must be checked as well.

Resolves: LXSWI9X1517-101
CRs-Fixed: 977563
Signed-off-by: Srinivasarao P <spathi@codeaurora.org>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I5fbec3722407d2f3d0ff0d9f7097c5889e31fd62