trac:4370 - Clean up USB audio gadget on restart
Sound PCM devices were not properly closed when the audio gadget
was disabled. As a result, attempting to restart the gadget would
cause the kernel to busy-loop in a schedule() call trying to open
a (previously unclosed) PCM device.

Once that was fixed, the ECM gadget would fail to restart, due to
running out of USB endpoints. This happened because the audio
gadget's .disable routine missed to disable and reclaim USB
endpoints.

The fix includes disabling/reclaiming USB endpoints and performing
an audio gadget cleanup in the gadget disable procedure.

patch_1: Drain capture/playback queue before disabling endpoint.

Change-Id: Ia2f7e3a2af72d310ddf9c488710a85ebfe1e66ac
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>