WPx5 - external PCM/I2S support
The WPx5 (per CF3 pin-out standard) has only one PCM bus. This
is used for the internal codec when present or the external
PCM/I2S bus for connecting to other audio hardware.
Initially, the WPx5 leveraged the audio configuration of the AR8,
which was using the same internal codec (WM8944), but was using
the secondary PCM bus for external PCM/I2S. This needed to be
changed to primary bus for WPx5.

 - Remove secondary PCM/I2S from CF3 platforms
     to free up GPIO 25-28 for other purposes (DEV86131)
 - Use primary PCM bus (similar to MC7) for external
     PCM/I2S modes.

 - Requires application changes included in ANO84833 (for AT!AV)

Resolves: trac:4901, ANO84833, DEV90619
Relates: trac:4783, DEV86131

Change-Id: I599c3ed71dadfeb497aecda75767623a5fbd595c
Signed-off-by: Sam Anderson <sanderson@sierrawireless.com>