Initialize PMIC GPIO 1 and 2 to avoid current leakage
PMIC GPIO 1 and 2 are designed to identify the RF module if different RF
modules are supported with single modem firmware release. The feature is
currently not used and the pins are grounded. PMIC GPIO 2 is defaulted
to analog output (1.8V), which incurs significant current leakage (~3mA).
Both GPIO pins must be configured as input pins with internal pull-down
to avoid the current leakage.

Change-Id: Idc1da4fb799a482fef34f8c1fc3460b0eb364e3e
Resolves: trac:4935
Signed-off-by: Dylan Zhang <dyzhang@sierrawireless.com>