Block extra MCU call on WPx5 powerdown
With the addition of the software initiated powerdown feature,
enabling ULPM triggers a call to powerdown, which, in turn, causes
another command to be sent to the MCU. By that time the MCU has
stopped communications, so the command fails generating several
kernel error messages before shutdown.

Solution:
If ULPM is already enabled (MDM will soon power down), don't send
another command to MCU to shutdown.

Resolves: LXSWI9X1517-100
Change-Id: I23a87b9fe31834fc9a1c8fbd867b773ef433eb80
Signed-off-by: Sam Anderson <sanderson@sierrawireless.com>