ipc_router: LINUX IPC Router Binding Any Port as a Control Port
IPC Router binds any port as a control port and moves it from the client
port list to control port list. Misbehaving clients can exploit this
incorrect behavior.

IPC Router to check if the port is a client port before binding it as a
control port.

Notes, D. Marinkovic
====================
This is modified version of patch located at
https://source.codeaurora.org/quic/la/kernel/msm-3.18/commit/?id=9e8bdd63f7011dff5523ea435433834b3702398d

Problem description is located at
https://www.codeaurora.org/projects/security-advisories/linux-ipc-router-binding-any-port-control-port-cve-2016-2059

CRs-Fixed: 974577
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Modified-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Resolves: Jira:LXSWI9X1517-111

Change-Id: I9f189b76967d5f85750218a7cb6537d187a69663