Initialize external GPIOs to input no pull
Root Cause
Several issues have been noted involving gpio2, 8, and 13 behaving
incorrectly. Gpio2 in particular does not function correctly in sysfs
when configured as an output, unless the pull type is modified first.
This behaviour seems to be related to the default configurations, which
are inconsistent between each gpio and in some cases interfere
with normal operation.

Solution
Initialize user gpios on startup.

Resolves: QTI9X1555-753
Change-Id: Ib3d3e9896aabb6c19f020d03cd843be4732d4215
Signed-off-by: Mitchell Nichols <MNichols@sierrawireless.com>