mdm9615: Defer claiming of IRQ GPIO for TI wifi chip
TI's WL18XX device is not part of any MDM9615 platform; it was
intended that a daughter card with device is added to the MangOH
platform. If the IRQ GPIO for this device is allocated at boot
time, then any other device in slot 0 of MangOH platform wouldn't
be able to claim it.

Defer allocation of this GPIO until a daughter card in slot 0 is
detected. GPIO will be claimed by a separate MangOH module that
performs this detection.

Change-Id: Ib220eef15e2888ab4608c6a0594de245c3bf20ec
Resolves:LXSWI9X1517-51
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>
(cherry picked from commit 6414f4bf2db97161ca7859218a901ed257bcdb21)