Smack: Fix smack file system mount problem
For some reason, change introduced by
'smack: pass error code through pointers' changeset is causing
ubifs file system not to mount, when SMACK options are one of the
parameters.

This is just temporary patch until I figure out what is going on.
Needless to say, this needs to be fixed properly.

Resolves: Jira:LXSWI9X1517-123
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: If6b7ecd48c9506311558fe287d358fee7e96b3f2