wm8944: add config option to enable interrupts
wm8944 uses msmgpio 86 as a parent interrupt line on AR hardware types, but
on WPx5 this line is actually an internal GPS_LNA_EN signal. This causes a
significant increase in boot time when gps auto start is enabled. The interrupt
signal from wm8944 is not routed on the mangOH board, so we can disable
interrupts by omitting the IRQ line from the wm8944 platform data. Add a config
option for wm8944 interrupts and disable by default.

Resolves: QTI9X1555-135
Change-Id: Idb559fd9ea5a41475d5d1a483b3bcf9acfc8e7a6
Signed-off-by: Mitchell Nichols <MNichols@sierrawireless.com>