Revert "Smack: Assign smack_known_web as default smk_in label ..."
This reverts commit aab24bcd4b15c63a22f116f129dcf11b2e3891fc

For some reason, the patch we are reverting is preventing NFS mount,
and disrupts development workflow. This is not the fix, it is here
to allow developers to continue their work. Once we have real fix in
place, this revert should be reverted. Only then related Jira ticket
should be closed.

Resolves: Jira:LXSWI9X1517-188
Reported-by: Yan Kong <ykong@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>

Change-Id: I5dfad26f4cc469074e7b15855409bb16948c1e0a