gpiolib: enable wakeup when setting up IRQ
Currently, GPIO interrupts cannot wake the system from sleep mode, which
may be due to the fact that the outdated gpio-v2 driver is used in this
kernel. Add a call to irq_set_irq_wake() in gpio_setup_irq() to allow GPIO
interrupts to wake the system from suspend state as per Qualcomm's
suggestion in case 03470156.

Resolves: QTI9X1555-1228
Change-Id: I41805da3e217819e4107d4225ca586af09b25dbe
Signed-off-by: Mitchell Nichols <MNichols@sierrawireless.com>